+++ src/main/java/com/vonage/client/verify/CheckRequest.java
@@ -16,28 +16,53 @@
 package com.vonage.client.verify;
 
 public class CheckRequest {
-    private final String requestId;
-    private final String code;
-    private final String ipAddress;
+    private final String requestId, code, ipAddress;
 
+    /**
+     * @param requestId The Verify request to check.
+     * This is the request_id you received in the response to the Verify request.
+     *
+     * @param code The verification code entered by your user. Between 4 and 6 characters.
+     */
     public CheckRequest(String requestId, String code) {
         this(requestId, code, null);
     }
 
+    /**
+     *
+     * @param ipAddress No longer used
+     * @deprecated Please use {@link CheckRequest#CheckRequest(String, String)}.
+     */
+    @Deprecated
     public CheckRequest(String requestId, String code, String ipAddress) {
-        this.requestId = requestId;
-        this.code = code;
+        if ((this.requestId = requestId) != null && requestId.length() > 32) {
+            throw new IllegalArgumentException("request_id '"+requestId+"' is longer than 32 characters");
+        }
+        if ((this.code = code) != null && (code.length() < 4 || code.length() > 6)) {
+            throw new IllegalArgumentException("code '"+code+"' is not between 4 and 6 characters long");
+        }
         this.ipAddress = ipAddress;
     }
 
+    /**
+     * @return The Verify request to check.
+     * This is the request_id you received in the response to the Verify request.
+     */
     public String getRequestId() {
         return requestId;
     }
 
+    /**
+     * @return The verification code entered by your user.
+     */
     public String getCode() {
         return code;
     }
 
+    /**
+     * @deprecated This field is no longer used.
+     */
+    @Deprecated
     public String getIpAddress() {
         return ipAddress;
     }